Added servo work. Zero trigger does not work yet

+++ src/Commands/OperatorInputShooter.h
@@ -8,7 +8,7 @@
 class OperatorInputShooter : public Command {
 private:
 	double shooterSpeed = 0;
-	int shooterServoPosition = 0;
+	double shooterServoPosition = 0;
 	std::shared_ptr<LumberJack> lumberJack;
 
 public:
